refactor(gsd-extension): move Projection bodies into WorktreeStateProjection
Slice 7 / step A of ADR-016. Bodies of projectRootToWorktree,
projectWorktreeToRoot, and finalizeProjectionForMerge now live in
worktree-state-projection.ts as the canonical implementations. Their
private helpers (isSamePath, forceOverwriteAssessmentsWithVerdict,
ROOT_DIAGNOSTIC_FILES) move with them.

The legacy syncProjectRootToWorktree, syncStateToProjectRoot, and
syncWorktreeStateBack exports in auto-worktree.ts are now thin
delegators to the Module's _*Impl exports, kept alive only for the
remaining slice-7 caller migrations. They are retired in step F.

The Module's deletion test now passes: deleting it scatters the
per-direction bug-hardened rules (#1886, #2184, #2478, #2821) across
every caller instead of inlining a one-line delegator.
